JWS: Add explicit, separate verify step
Instead of always assuming we can verify with a key at deserialization
time, allow a key to be None and perform verification as a separate step.

The code reads better and we can easily try multiple keys after
deserialization is performed once. this also gets rid of the paramter
to raise on error, as a None key takes the role of not raising on
verification errors (still raises on format errors, but that is fine.

Signed-off-by: Simo Sorce <simo@redhat.com>