Provide a better way to export public keys
This makes the same call export just the public keys if private_key is
set to False.

In hindsight if we had this from the start the default would be to export
public keys and you'd have to actively pass a True value to export private
ones, but I do not want to break applications now.

Not only this API it is somewhat more natural (with the above caveat), it also
allows to do exactlyu the same for JWK Sets with substantially no code
changes.

Signed-off-by: Simo Sorce <simo@redhat.com>