Turn JWKSet into a special dictionary
The JWKSet object is now a class derived from dict, with a special element
'keys' that is forced to be a class derived by set which enforces only JWK
object instances as elements.

A new classmethod "from_json" is added to create a JWKset out of a json object
directly.

The interface is fully backward compatible with the existing APIs otherwise.

Signed-off-by: Simo Sorce <simo@redhat.com>
Reviewed-by: Nathaniel McCallum <npmccallum@redhat.com>
Closes #31
Closes #33