The Pbes2 algorithm accepts only bytes or string
The Pbes2 algorithm is supposed to get a low entropy password so we break
the interface so that JWK keys are not accepted at all.

This is in order to avoid potential confusion in users where a low entropy
password is coerced into an "oct" key and then potentially used as a high
entropy symmetric key with other algorithms yielding very poor security.

By making password mutually eclusive with keys we try to avoid this mistakes
by users.

Signed-off-by: Simo Sorce <simo@redhat.com>
Closes #40