Fix encoding length of EC keys Coordinates
According to RFC 7518 6.2.1.2 the Coordinates need to be encoded such
that the size of the encoded string is fixed (based on private curve
key_size). Fix _encode_int to take in account the bit_size given
python-cryptography gives us python integers and that may lead to
truncation of leading zeros when doing the conversion via hex().

Fixes: #158

Signed-off-by: Simo Sorce <simo@redhat.com>