Make sure an empty dict is a valid JWT payload
Nowhere in the spec it says the payload must not be an empty dictionary.
It makes little sense to not have any claims, because then, what do you
claim? But that's a problem for the user, especially given the fact you
could already pas in the literal string '{}' and it would get signed.

However a completely empty string is invalid, at least one character,
even just a space is needed.

Fixes #187

Signed-off-by: Simo Sorce <simo@redhat.com>