Deprecate RSA1_5 and remove from defaults
RSA1_5 relies on the PKCS1 v 1.5 padding algorithm.

This algorithm has been shown time and again to be very difficult to
implement correctly because it requires a fully constant time
implementation including in application level error handling.

Python Cryptography which is our crypto engine documents that the
current API is not constant time.

This patch deprecates the algorithm and removes it from the default
allowed algorithms list. This means that code needs to be changed in
order to explicitly allow RSA1_5 going forward.

Signed-off-by: Simo Sorce <simo@redhat.com>